add env support and valibot example usage on backend app

+++ backend/src/meals/meals.service.ts
@@ -1,7 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Meal } from './meal.entity';
+import { minValue, number, safeParse } from 'valibot';
+
+const MealIdSchema = number([minValue(1)]);
 
 @Injectable()
 export class MealsService {
@@ -23,6 +26,12 @@ export class MealsService {
   }
 
   async delete(mealId: number) {
-    return this.mealsRepository.delete(mealId);
+    const validation = safeParse(MealIdSchema, mealId);
+
+    if (validation.success) {
+      return this.mealsRepository.delete(mealId);
+    }
+
+    throw new BadRequestException();
   }
 }
